+++ Modules/antibodies/Controller/AntibodiesconfigController.php
@@ -41,7 +41,7 @@ class AntibodiesconfigController extends CoresecureController
         $lang = $this->getLanguage();
 
         // maintenance form
-        $formMenusactivation = $this->menusactivationForm($id_space, 'antibodies', $lang);
+        $formMenusactivation = $this->menusactivationForm($id_space, 'antibodies', $lang, CoreSpace::$MANAGER);
         if ($formMenusactivation->check()) {
             $this->menusactivation($id_space, 'antibodies', 'person');
             return $this->redirect("antibodiesconfig/".$id_space);