+++ Modules/mailer/View/Mailer/indexAction.php
@@ -41,9 +41,9 @@
                 <label for="to" class="control-label col-12 col-md-2"><?php echo MailerTranslator::To($lang) ?></label>
                 <div class="col-12 col-md-10">
                     <select class="form-control" id="to" name="to">
-                        <OPTION value="all" > all </OPTION>
+                        <OPTION value="all" > <?php echo CoreTranslator::Users($lang) ?> </OPTION>
                         <OPTION value="managers" > <?php echo CoreTranslator::Managers($lang) ?> </OPTION>
-                        <?php if($superAdmin) { ?><OPTION value="admins" > <?php echo CoreTranslator::Admins($lang) ?> </OPTION><?php } ?>
+                        <?php if ($superAdmin) { ?><OPTION value="admins" > <?php echo CoreTranslator::Admins($lang) ?> </OPTION><?php } ?>
                         <?php foreach ($areasList as $area): ?>
                             <?php
                             $areaname = $this->clean($area['name']);
@@ -76,8 +76,7 @@
             <div class="form-group row">
                 <label for="content" class="control-label col-12 col-md-2"><?php echo MailerTranslator::Content($lang) ?></label>
                 <div class="col-12 col-md-10">
-                    <textarea class="form-control" id="content" name="content">
-                    </textarea>
+                    <textarea rows="8" class="form-control" id="content" name="content"></textarea>
                 </div>
             </div>
             <br><br/>
@@ -116,7 +115,7 @@
                 <td><?php echo substr(trim($mail['message']), 0, 15); ?>...</td>
                 <?php if($role >= $editRole) { ?>
                 <td><?php echo MailerTranslator::dest($mail['type'], $lang); ?></td>
-                <td><a href="mailer/<?php echo $id_space; ?>/delete/<?php echo $mail['id']; ?>"><button type="button" class="btn"><?php echo CoreTranslator::Delete($lang) ?></button></a>
+                <td><a href="mailer/<?php echo $id_space; ?>/delete/<?php echo $mail['id']; ?>"><button type="button" class="btn btn-danger"><?php echo CoreTranslator::Delete($lang) ?></button></a>
                 <?php } ?>
             </tr>
         <?php } ?>