+++ Modules/services/Controller/ServicesconfigController.php
@@ -40,7 +40,7 @@ class ServicesconfigController extends CoresecureController
         $modelCoreConfig = new CoreConfig();
 
         // maintenance form
-        $formMenusactivation = $this->menusactivationForm($id_space, 'services', $lang);
+        $formMenusactivation = $this->menusactivationForm($id_space, 'services', $lang, CoreSpace::$MANAGER);
         if ($formMenusactivation->check()) {
             $this->menusactivation($id_space, 'services', 'basket');
             return $this->redirect("servicesconfig/" . $id_space);