+++ Modules/services/Model/SeProject.php
@@ -631,10 +631,7 @@ class SeProject extends Model
         $details = explode(";", $invoiceItem["details"]);
         if (count($details) < 2) {
             Configuration::getLogger()->warning('[services][getInfoFromInvoice] Invalid invoice item', ['item' => $invoiceItem]);
-            $info = array();
-            $info['closed_by'] = "";
-            $info['closed_by_in'] = "";
-            return $info;
+            return null;
         }
         $proj = explode("=", $details[count($details) - 2]);
         $projUrl = explode("/", $proj[1]);
@@ -644,9 +641,7 @@ class SeProject extends Model
         if ($req->rowCount() > 0) {
             $info = $req->fetch();
         } else {
-            $info = array();
-            $info['closed_by'] = "";
-            $info['closed_by_in'] = "";
+            return null;
         }
 
         if ($info['closed_by']) {
@@ -839,7 +834,7 @@ class SeProject extends Model
         for ($i = 0; $i < count($data); $i++) {
             $clientInfo = $modelClient->get($id_space, $data[$i]['id_resp']);
 
-            $data[$i]['resp'] = $clientInfo["contact_name"];
+            $data[$i]['resp'] = $clientInfo ? $clientInfo["contact_name"]: 'unknown';
             $data[$i]['user'] = $modelUser->getUserFUllName($data[$i]['id_user']);
             $data[$i]['unit'] = $modelClient->getInstitution($id_space, $data[$i]['id_resp']);
             $data[$i]["sample_cabinet"] = $modelSampleCabinet->getFullName($id_space, $data[$i]["id_sample_cabinet"]);